Updated print statements to be more descriptive. Changed the nandleKilledQueueInsertion to be checking the interrupted PCB, not the running PCB. Ran many tests to find out what's wrong with the program. Some things I've found:
We make too many PCBs. If we make the MAKE_PCB_CHANCE_DOMAIN too large though we experience a lot of deadlock. I will be doing more tests to find out what the maximum value we can go to is.

Next. It seems that at certain times when a context switch is happening to move an io Type into the blocked queue, its PC value will still be incremented. This isn't a huge deal for io, but if that happens with Mutex lock/unlock and the wait/signal then we're going to have a problem.

Also, I'm not sure the Deadlock monitor is working properly. Tests are currently going on to be sure it is doing what is required, but it doesn't seem like it accurately detects deadlock in every case. Of course, it's only meant to be an approximation, but it's something to look into.

Lastly, there needs to be some refactoring of the output. Along with descriptive names for the PCBs, Mutexes, and ConditionVariables, the output needs to be more robust. Especially when actually incrementing the PC for a process. There should be a single note saying that the Program is running, but should only print that out once after any major event. We don't need 500 lines in the log file saying it for a single PCB at a time.